refactor: extract 6 small blueprints to routes/meta.py
Phase 5.12 of incremental modularisation. Bundles six tiny Blueprints
(13 routes total) into one file because each is small (1-3 routes) and
they're all auth/meta/observability plumbing — keeps CI churn minimal.

  bp_version        (2)  /api/version, /api/update
  bp_gateway        (3)  /api/gw/{config,invoke,rpc}
  bp_auth           (3)  /api/auth/check, /auth, /  (main page)
  bp_otel           (3)  /v1/metrics, /v1/traces, /api/otel-status
  bp_version_impact (1)  /api/version-impact
  bp_clusters       (1)  /api/clusters

Module-level helpers, state (GATEWAY_URL, GATEWAY_TOKEN, _ws_client,
_ws_connected, _pypi_cache, metrics_store, DASHBOARD_HTML, etc.) and
the @app.before_request _check_auth hook stay in dashboard.py and are
reached from routes/meta.py via late `import dashboard as _d`.

dashboard.py: -443 lines.
